end_to_end_demo/server_python/requirements.txt
click==8.1.3
Flask==2.1.1
Flask-Cors==3.0.10
grpcio==1.44.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
protobuf==3.20.1
six==1.16.0
userwatch-python==0.0.3
Werkzeug==2.1.2
zipp==3.8.0
